+++ packages/ui-components/web/src/Components/SettingsAddItem.tsx
@@ -1,20 +1,63 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import type { CSSProperties } from 'react';
 import React from 'react';
+import type { CSSProperties } from 'react';
 import FileInput from './FileInput';
-import PlusIcon from '../Icons/PlusIcon';
+import Image from './Image';
+import Tooltip from 'wix-rich-content-common/libs/Tooltip';
+import { PlusIcon, ReplaceIcon, DeleteIcon } from '../Icons';
 import styles from '../../statics/styles/settings-add-item.scss';
 import classNames from 'classnames';
-import type { RichContentTheme } from 'wix-rich-content-common';
+import type { RichContentTheme, TranslationFunction } from 'wix-rich-content-common';
+
+interface AddItemOverlayProps {
+  handleReplace?: () => void;
+  handleDelete?: () => void;
+  t: TranslationFunction;
+  src: string;
+  theme?: RichContentTheme;
+  alt?: string;
+}
+
+const AddItemOverlay: React.FC<AddItemOverlayProps> = ({
+  handleReplace,
+  handleDelete,
+  src,
+  theme,
+  alt,
+  t,
+}) => (
+  <div className={styles.settingsAddItem_overlay}>
+    <div className={styles.settingsAddItem_overlay_icons}>
+      <Tooltip content={t('Settings_Replace_Image_Tooltip')} place="top">
+        <ReplaceIcon tabIndex={0} onClick={handleReplace} />
+      </Tooltip>
+      <Tooltip content={t('Settings_Remove_Image_Tooltip')} place="top">
+        <DeleteIcon tabIndex={0} onClick={handleDelete} />
+      </Tooltip>
+    </div>
+    <Image
+      className={styles.settingsAddItem_overlay_image}
+      src={src}
+      resizeMode={'cover'}
+      theme={theme || styles}
+      alt={alt}
+      t={t}
+    />
+  </div>
+);
 
 interface SettingsAddItemProps {
   handleFileChange: (files?: any, itemPos?: any) => void;
   isMobile?: boolean;
   handleFileSelection?: () => void;
+  handleDelete?: () => void;
   inlineStyle?: CSSProperties;
   theme?: RichContentTheme;
+  t: TranslationFunction;
+  src?: string;
+  alt?: string;
   uploadMediaLabel: string;
-  accept: string;
+  accept?: string;
   dataHook: string;
 }
 
@@ -27,23 +70,40 @@ const SettingsAddItem: React.FC<SettingsAddItemProps> = ({
   uploadMediaLabel,
   accept,
   dataHook,
+  handleDelete,
+  src,
+  alt,
+  t,
 }) => {
   return (
-    <FileInput
-      className={classNames(styles.settingsAddItem, styles.filesItem, {
-        [styles.mobile]: isMobile,
-      })}
-      dataHook={dataHook}
-      onChange={handleFileChange}
-      handleFileSelection={handleFileSelection}
-      multiple
-      theme={theme}
-      title={uploadMediaLabel}
-      style={inlineStyle}
-      accept={accept}
-    >
-      <PlusIcon />
-    </FileInput>
+    <div className={styles.settingsAddItem_container}>
+      <FileInput
+        className={classNames(styles.settingsAddItem, styles.filesItem, {
+          [styles.mobile]: isMobile,
+        })}
+        dataHook={dataHook}
+        onChange={handleFileChange}
+        handleFileSelection={handleFileSelection}
+        tabIndex={src ? -1 : 0}
+        multiple
+        theme={theme}
+        title={uploadMediaLabel}
+        style={inlineStyle}
+        accept={accept}
+      >
+        <PlusIcon />
+      </FileInput>
+      {src && (
+        <AddItemOverlay
+          handleDelete={handleDelete}
+          handleReplace={handleFileSelection}
+          theme={theme}
+          src={src}
+          alt={alt}
+          t={t}
+        />
+      )}
+    </div>
   );
 };